% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENSUP-package.R
\docType{package}
\name{CENSUP-package}
\alias{CENSUP}
\alias{CENSUP-package}
\title{CENSUP: Standartization of Brazilian higher education census data}
\description{
Cleans and standardises data from the Brazilian higher education census, to allow for easy panel analysis of the available variables.
}
\author{
\strong{Maintainer}: João Gabriel Vasconcellos Godoy \email{jey.snow.dev@gmail.com}

Authors:
\itemize{
  \item João Gabriel Vasconcellos Godoy \email{jey.snow.dev@gmail.com}
}

}
\keyword{internal}
