% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_faculty.R
\name{clean_FAC}
\alias{clean_FAC}
\title{Read a csv file from CENSUP and create a standardized file for faculty level data}
\usage{
clean_FAC(READ_FROM, OUTPUT_TO = "RETURN", LANGUAGE = "ENG", CITY_NAMES = T)
}
\arguments{
\item{READ_FROM}{the path to a csv file or a folder containing csv files}

\item{OUTPUT_TO}{the path to a folder where the standardized file is to be written to.
Defaults to "return", whereby the function simply returns and doesn't save the table}

\item{LANGUAGE}{the language in which the data should be standardized to. defaults to English}

\item{CITY_NAMES}{convert city codes to names, TRUE by default, but slower}
}
\value{
if OUTPUT_TO = "RETURN", returns a data table or a list of data tables
}
\description{
Read a csv file from CENSUP and create a standardized file for faculty level data
}
\examples{
\dontrun{clean_FAC()}
}
