% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_HEI_FAC.R
\name{clean_HEI_FAC}
\alias{clean_HEI_FAC}
\title{Title}
\usage{
clean_HEI_FAC(
  READ_FROM_HEI,
  OUTPUT_TO = "RETURN",
  READ_FROM_FAC = "NONE",
  LANGUAGE = "ENG",
  CITY_NAMES = TRUE
)
}
\arguments{
\item{OUTPUT_TO}{the path to a folder where the standardized file is to be written to.
Defaults to "return", whereby the function simply returns and doesn't save the table}

\item{READ_FROM_FAC}{the path to a csv file or a folder containing csv files for faculty data}

\item{LANGUAGE}{the language in which the data should be standardized to. defaults to English}

\item{CITY_NAMES}{convert city codes to names, TRUE by default, but slower}

\item{READ_FROM}{the path to a csv file or a folder containing csv files for HEI data}
}
\value{
if OUTPUT_TO = "RETURN", returns a data table or a list of data tables
}
\description{
Title
}
\examples{
\dontrun{clean_HEI_FAC()}
}
