% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_HEI_FAC.R
\name{combine_HEI_FAC}
\alias{combine_HEI_FAC}
\title{Combines faculty data at the HEI level and merges with HEI data}
\usage{
combine_HEI_FAC(
  CLEAN_HEI,
  CLEAN_FAC,
  OUTPUT_TO = "RETURN",
  LANGUAGE = "ENG",
  CITY_NAMES = TRUE
)
}
\arguments{
\item{CLEAN_HEI}{A data.table with clean HEI data, or a list of such,
or a path to a file containing clean HEI data, or vector of such,
or a path to a folder containing only files with clean HEI data}

\item{CLEAN_FAC}{A data.table with clean FAC data, or a list of such,
or a path to a file containing clean FAC data, or vector of such,
or a path to a folder containing only files with clean FAC data}

\item{OUTPUT_TO}{the path to a folder where the standardized file is to be written to.
Defaults to "return", whereby the function simply returns and doesn't save the table}

\item{LANGUAGE}{the language in which the data should be standardized to. defaults to English}

\item{CITY_NAMES}{convert city codes to names, TRUE by default, but slower}
}
\value{
if OUTPUT_TO = "RETURN", returns a data table or a list of data tables
}
\description{
Combines faculty data at the HEI level and merges with HEI data
}
\examples{
\dontrun{combine_HEI_FAC()}
}
